Update server paths and add config file
