+++ server/src/util/config.js
@@ -1,0 +1,7 @@
+require("dotenv").config();
+
+const serverConfig = {
+  PORT: process.env.PORT || 8000,
+};
+
+module.exports = serverConfig;
