+++ server/src/util/encodeBody.js
@@ -1,0 +1,7 @@
+const encodeBody = (body) => {
+  return Object.keys(body)
+    .map((key) => encodeURIComponent(key) + "=" + encodeURIComponent(body[key]))
+    .join("&");
+};
+
+module.exports = encodeBody;
